+++ src/init.c
@@ -176,6 +176,7 @@ static const R_CallMethodDef callMethods[] = {
   {"rnng_version", (DL_FUNC) &rnng_version, 0},
   {"rnng_wait_thread_create", (DL_FUNC) &rnng_wait_thread_create, 1},
   {"rnng_write_cert", (DL_FUNC) &rnng_write_cert, 2},
+  {"rnng_writec", (DL_FUNC) &rnng_writec, 1},
   {NULL, NULL, 0}
 };